requirements.txt
flask
pillow
qrcode
requests
numpy
pandas
opencv-python
imageio
matplotlib
colormap
easydev